Refactored image initialization to Picture Repository along with Metadata extraction
